[refactor] token config package  rename and move
