+++ src/main/java/toy/bookchat/bookchat/security/token/openid/keys/GooglePublicKeys.java
@@ -1,9 +1,8 @@
 package toy.bookchat.bookchat.security.token.openid.keys;
 
-import toy.bookchat.bookchat.config.openid.OAuthPublicKey;
-
 import java.security.Key;
 import java.security.KeyFactory;
+import toy.bookchat.bookchat.config.token.openid.OAuthPublicKey;
 
 public class GooglePublicKeys implements OAuthPublicKey {
